Conflicts:
	src/freenet/node/RequestTracker.java

+++ src/freenet/io/comm/MessageFilter.java
@@ -53,7 +53,7 @@ public final class MessageFilter {
     /** If true, timeouts are relative to the start of waiting, if false, they are relative to
      * the time of calling setTimeout() */
     private boolean _timeoutFromWait;
-    private int _initialTimeout;
+    private long _initialTimeout;
     private MessageFilter _or;
     private Message _message;
     private long _oldBootID;
@@ -106,7 +106,7 @@ public final class MessageFilter {
      * @param timeout The time before this filter expires in ms
      * @return This message filter
      */
-	public MessageFilter setTimeout(int timeout) {
+	public MessageFilter setTimeout(long timeout) {
 		_setTimeout = true;
 		_initialTimeout = timeout;
 		_timeout = System.currentTimeMillis() + timeout;
@@ -286,7 +286,7 @@ public final class MessageFilter {
         notifyAll();
     }
 
-    public int getInitialTimeout() {
+    public long getInitialTimeout() {
         return _initialTimeout;
     }
